+++ BililiveRecorder.Core/SimpleWebhook/RecordEndData.cs
@@ -22,7 +22,7 @@ namespace BililiveRecorder.Core.SimpleWebhook
 
         public Guid EventRandomId { get; set; } = Guid.NewGuid();
 
-        public int RoomId { get; set; } = 0;
+        public long RoomId { get; set; } = 0;
         public string Name { get; set; } = string.Empty;
         public string Title { get; set; } = string.Empty;
         public string RelativePath { get; set; } = string.Empty;
